backend/foodgram/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.6
django-filter==22.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
exceptiongroup==1.1.0
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyJWT==2.6.0
pytest==6.2.5
pytest-django==4.5.2
pytest-pythonpath==0.7.4
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.14
